react-calendar 적용(날짜별 todoList

+++ src/App.js
@@ -1,9 +1,13 @@
 import React, { useEffect, useRef, useState } from 'react';
+import Calendar from 'react-calendar';
+import 'react-calendar/dist/Calendar.css'
+import moment from 'moment'
 import './App.css'
 
 export function App() {
   const [todoList, setTodoList] = useState([]);
   const [sequance, setSequance] = useState(null);
+  const [chooseDate, setChooseDate] = useState();
   const refTodoItem = useRef();
 
   useEffect(() => { //로컬 스토리지에 미리 sequance, todoList 삽입
@@ -19,9 +23,9 @@ export function App() {
     };
 
     let todo = JSON.parse(window.localStorage.getItem("todoList") ?? handleSetInit());
-
     setTodoList(todo);
     setSequance(Number(sequance));
+    setChooseDate(new Date())
   }, []);
 
   const handleTodoAdd = (item) => {
@@ -31,9 +35,8 @@ export function App() {
     if (item === '') {
       return;
     }
-
     let todo = [...todoList];
-    todo.push({ tf: false, id: sequance + 1, text: item });
+    todo.push({ chooseDate: moment(chooseDate).format('YYYYMMDD'), tf: false, id: sequance + 1, text: item });
 
     window.localStorage.setItem("todoList", JSON.stringify(todo));
     window.localStorage.setItem("sequance", String(sequance + 1));
@@ -59,12 +62,27 @@ export function App() {
     setTodoList(todo);
 
   };
+
+  const handleChooseDate = (val) => {
+    let date = moment(val).format('YYYYMMDD');
+    console.log(date)
+    setChooseDate(date);
+
+  }
   return (
     <div className='mainLayout'>
       <div className='todoLayout'>
         <div className='todoTop'>
           <div className='todoTitle'>
             To Do List
+          </div>
+          <div className='todoCalendar'>
+            <Calendar
+              locale='ko'
+              formatDay={(locale, date) => moment(date).format('D')}
+              onChange={(value) => handleChooseDate(value)}
+              value={chooseDate}
+            />
           </div>
           <div className='todoAdd'>
             <input
@@ -75,22 +93,27 @@ export function App() {
               +
             </div>
           </div>
+          {
+            todoList.map((item, i) => {
+              if (item.chooseDate === moment(chooseDate).format('YYYYMMDD')) {
+                return (
+                  <div className='todoItem' key={i}>
+                    <div className='todoCheckBox' onClick={() => handleTodoCheck(item.tf, i)}>
+                      <div className='checkIcon'>
+                        {item.tf ? '✔️' : ''}
+                      </div>
+                      <span>{item.text}</span>
+                    </div>
+                    <div className='deleteItem' onClick={() => handleTodoDelete(item.id)}>
+                      ✖️
+                    </div>
+                  </div>
+                )
+              }
+            }
+            )
+          }
         </div>
-        {
-          todoList.map((item, i) =>
-            <div className='todoItem' key={i}>
-              <div className='todoCheckBox' onClick={() => handleTodoCheck(item.tf, i)}>
-                <div className='checkIcon'>
-                  {item.tf ? '✔️' : ''}
-                </div>
-                <span>{item.text}</span>
-              </div>
-              <div className='deleteItem' onClick={() => handleTodoDelete(item.id)}>
-                ✖️
-              </div>
-            </div>
-          )
-        }
       </div>
     </div>
   );
